Fix/product types 2

+++ src/main/java/com/develhope/spring/services/ProductTypeService.java
@@ -1,9 +1,11 @@
 package com.develhope.spring.services;
 
+import com.develhope.spring.daos.ProductDao;
 import com.develhope.spring.exceptions.InvalidProductTypeException;
 import com.develhope.spring.models.ResponseCode;
 import com.develhope.spring.models.ResponseModel;
 import com.develhope.spring.models.dtos.ProductTypeDto;
+import com.develhope.spring.models.entities.ProductEntity;
 import com.develhope.spring.models.entities.ProductTypeEntity;
 import com.develhope.spring.mappers.ProductTypeMapper;
 import com.develhope.spring.daos.ProductTypeDao;
@@ -13,7 +15,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductTypeService {
@@ -21,12 +22,14 @@ public class ProductTypeService {
     private final ProductTypeDao productTypeDao;
     private final ProductTypeMapper productTypeMapper;
     private final ProductTypeValidator productTypeValidator;
+    private final ProductDao productDao;
 
     @Autowired
-    public ProductTypeService(ProductTypeDao productTypeDao, ProductTypeMapper productTypeMapper, ProductTypeValidator productTypeValidator) {
+    public ProductTypeService(ProductTypeDao productTypeDao, ProductTypeMapper productTypeMapper, ProductTypeValidator productTypeValidator, ProductDao productDao) {
         this.productTypeDao = productTypeDao;
         this.productTypeMapper = productTypeMapper;
         this.productTypeValidator = productTypeValidator;
+        this.productDao = productDao;
     }
 
     /**
@@ -70,22 +73,45 @@ public class ProductTypeService {
     }
 
     /**
-     * @effect productType is an identifier, can't be updated.
-     * This method replaces the productType to be upgraded with a new one
      * @param productType        productType name
      * @param productTypeUpdates the new productType that replace the old one
      * @return a productType updated
+     * @effect productType is an identifier, can't be updated.
+     * This method replaces the productType to be upgraded with a new one
      */
     public ResponseModel updateProductType(String productType, String productTypeUpdates) {
         ProductTypeEntity productTypeToUpdate = this.productTypeDao.findByProductType(productType);
+        List<ProductEntity> products = productDao.findByProductTypesContains(productTypeToUpdate);
+
         if (productTypeToUpdate == null) {
             return new ResponseModel(ResponseCode.D).addMessageDetails("ProductType not found");
+        }
+
+        //if a productType is assigned to a product or a list of products
+        if (productTypeUpdates != null && !products.isEmpty()) {
+            for (ProductEntity p : products) {
+                p.getProductTypes().remove(productTypeToUpdate);
+                productDao.save(p);
+            }
+
+            this.productTypeDao.delete(productTypeToUpdate);
+            ProductTypeEntity newProductType = new ProductTypeEntity();
+            newProductType.setProductType(productTypeUpdates);
+            productTypeDao.save(newProductType);
+
+            for (ProductEntity p : products) {
+                p.getProductTypes().add(newProductType);
+                productDao.save(p);
+            }
+            return new ResponseModel(ResponseCode.G, productTypeMapper.toDto(this.productTypeDao.saveAndFlush(newProductType)));
+
         } else if (productTypeUpdates != null) {
             this.productTypeDao.delete(productTypeToUpdate);
             ProductTypeEntity newProductType = new ProductTypeEntity();
             newProductType.setProductType(productTypeUpdates);
             return new ResponseModel(ResponseCode.G, productTypeMapper.toDto(this.productTypeDao.saveAndFlush(newProductType)));
         }
+
         return new ResponseModel(ResponseCode.A).addMessageDetails("Impossible to update, the body should not be null");
     }
 
@@ -97,16 +123,43 @@ public class ProductTypeService {
     public ResponseModel deleteProductType(String productTypeName) {
         if (this.productTypeDao.findByProductType(productTypeName) == null) {
             return new ResponseModel(ResponseCode.D).addMessageDetails("ProductType not found");
-        } else {
+        }
+
+        //if a productType is assigned to a product or a list of products
+        ProductTypeEntity productTypeToDelete = this.productTypeDao.findByProductType(productTypeName);
+        List<ProductEntity> products = productDao.findByProductTypesContains(productTypeToDelete);
+        if (productTypeName != null && !products.isEmpty()) {
+            for (ProductEntity p : products) {
+                p.getProductTypes().remove(productTypeToDelete);
+                this.productDao.save(p);
+            }
+
             productTypeDao.deleteByProductType(productTypeName);
             return new ResponseModel(ResponseCode.H).addMessageDetails("ProductType successfully deleted");
+
         }
+
+        productTypeDao.deleteByProductType(productTypeName);
+        return new ResponseModel(ResponseCode.H).addMessageDetails("ProductType successfully deleted");
     }
 
     /**
      * Delete all product types
      */
     public ResponseModel deleteAllProductTypes() {
+        List<ProductTypeEntity> allProductTypes = this.productTypeDao.findAll();
+        List<ProductEntity> allProducts = this.productDao.findAll();
+
+        if (!allProducts.isEmpty()) {
+            for (ProductEntity p : allProducts) {
+                p.getProductTypes().removeAll(allProductTypes);
+                this.productDao.save(p);
+            }
+
+            this.productTypeDao.deleteAll();
+            return new ResponseModel(ResponseCode.H).addMessageDetails("All productTypes have been deleted");
+        }
+
         this.productTypeDao.deleteAll();
         return new ResponseModel(ResponseCode.H).addMessageDetails("All productTypes have been deleted");
     }
